add support for MySQL preparedBatch
