+++ vertx-mysql-client/src/main/java/io/vertx/mysqlclient/impl/MySQLConnectionImpl.java
@@ -15,16 +15,9 @@ import io.vertx.mysqlclient.impl.command.PingCommand;
 import io.vertx.mysqlclient.impl.command.ResetConnectionCommand;
 import io.vertx.mysqlclient.impl.command.SetOptionCommand;
 import io.vertx.mysqlclient.impl.command.StatisticsCommand;
-import io.vertx.sqlclient.Row;
-import io.vertx.sqlclient.RowSet;
-import io.vertx.sqlclient.SqlResult;
 import io.vertx.sqlclient.Transaction;
-import io.vertx.sqlclient.Tuple;
 import io.vertx.sqlclient.impl.Connection;
 import io.vertx.sqlclient.impl.SqlConnectionImpl;
-
-import java.util.List;
-import java.util.stream.Collector;
 
 public class MySQLConnectionImpl extends SqlConnectionImpl<MySQLConnectionImpl> implements MySQLConnection {
 
@@ -70,16 +63,6 @@ public class MySQLConnectionImpl extends SqlConnectionImpl<MySQLConnectionImpl> 
   @Override
   public Transaction begin(boolean closeOnEnd) {
     throw new UnsupportedOperationException("Transaction is not supported for now");
-  }
-
-  @Override
-  public MySQLConnectionImpl preparedBatch(String sql, List<Tuple> batch, Handler<AsyncResult<RowSet>> handler) {
-    throw new UnsupportedOperationException("PreparedBatch is not supported for now");
-  }
-
-  @Override
-  public <R> MySQLConnectionImpl preparedBatch(String sql, List<Tuple> batch, Collector<Row, ?, R> collector, Handler<AsyncResult<SqlResult<R>>> handler) {
-    throw new UnsupportedOperationException("PreparedBatch is not supported for now");
   }
 
   @Override
